Compiler warnings - type conversion - gtest

+++ test/gtest-1.7.0/include/gtest/gtest.h
@@ -1498,7 +1498,7 @@ AssertionResult CmpHelperEQ(const char* expected_expression,
 								 // signed/unsigned mismatch.
 #endif
 
-	if (expected == actual)
+	if ((int)expected == actual)
 	{
 		return AssertionSuccess();
 	}
